requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-dotenv
openai
requests
httpx
langgraph
langchain
langchain-core
langchain-community
langchain-groq
groq
sentence-transformers
chromadb
pytest
anyio
typing_extensions